+++ include/gear/gearxml/tinyxml.h
