+++ src/gearxml/CalorimeterParametersXML.cc
@@ -3,7 +3,7 @@
 #include "gearxml/XMLHandlerMgr.h"
 #include "gearxml/GearParametersXML.h"
 
-#include "tinyxml/tinyxml.h"
+#include "gearxml/tinyxml.h"
 #include "gearimpl/CalorimeterParametersImpl.h"
 // #include "gear/LayerLayout.h"
 #include "gear/GearMgr.h"
