+++ src/gearxml/FTDParametersXML.cc
@@ -4,7 +4,7 @@
 #include "gearxml/XMLHandlerMgr.h"
 #include "gearxml/GearParametersXML.h"
 
-#include "tinyxml/tinyxml.h"
+#include "gearxml/tinyxml.h"
 #include "gearimpl/FTDParametersImpl.h"
 
 #include "gear/GearMgr.h"
