+++ src/gearxml/VersatileDiskRowLayoutXML.cc
@@ -3,7 +3,7 @@
 #include "gearxml/XMLHandler.h"
 // #include "gearxml/GearParametersXML.h"
 
-#include "tinyxml/tinyxml.h"
+#include "gearxml/tinyxml.h"
 #include "gearimpl/VersatileDiskRowLayout.h"
 #include "gear/GearMgr.h"
