+++ src/gearxml/tinyxmlparser.cc
@@ -22,12 +22,12 @@ must not be misrepresented as being the original software.
 distribution.
 
  F.Gaede, DESY : changed extension to .cc  for use with gear
-                 and include from "tinyxml/tinyxml.h"
+                 and include from "gearxml/tinyxml.h"
                : put in namespace gear
    $Id: tinyxmlparser.cc,v 1.2 2008-12-19 13:52:34 gaede Exp $
 */
 
-#include "tinyxml/tinyxml.h"
+#include "gearxml/tinyxml.h"
 #include <ctype.h>
 #include <stddef.h>
