+++ src/test/mergeXML.cc
@@ -1,4 +1,4 @@
-#include "tinyxml/tinyxml.h"
+#include "gearxml/tinyxml.h"
 #include "gearxml/MergeXML.h"
 #include "gear/GEAR.h"
